requirements.txt
fastapi==0.109.2
uvicorn==0.25.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
pytest==7.4.3
pytest-asyncio==0.21.1